Download fio output from zonal buckets using gcsfuse
This is bypass the problem that gcloud storage cp
is not supported for zonal buckets yet.
For this support, a new environment var option
`zonal` has been introduced. When `zonal` is passed
as true, the outputs from fio workloads are downloaded
using gcsfuse rather than gcloud. By default, `zonal`
is false. It should be set to true of even one of
your fio workloaded uses a zonal bucket.